+++ src/ordered_collection_common.mli
@@ -2,7 +2,7 @@
 
 open! Import
 
-(** [get_pos_len], [get_pos_len_exn], and [validate_pos_len_exn] are intended to be used
+(** [get_pos_len], [get_pos_len_exn], and [check_pos_len_exn] are intended to be used
     by functions that take a sequence (array, string, bigstring, ...) and an optional
     [pos] and [len] specifying a subrange of the sequence.  Such functions should call
     [get_pos_len] with the length of the sequence and the optional [pos] and [len], and it
@@ -25,7 +25,7 @@ open! Import
       len' = match len with None -> length - pos | Some i -> i
     v} *)
 
-val get_pos_len     : ?pos:int -> ?len:int -> length:int -> (int * int, string) Result.t
+val get_pos_len     : ?pos:int -> ?len:int -> length:int -> (int * int) Or_error.t
 val get_pos_len_exn : ?pos:int -> ?len:int -> length:int -> int * int
 
 (** [check_pos_len_exn ~pos ~len ~length] raises unless [pos >= 0 && len >= 0 && pos + len